requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.7.0
bleach==6.0.0
brotlipy==0.7.0
build==0.10.0
bump2version==1.0.1
bumpversion==0.6.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.4
Cerberus==1.3.5
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
crashtest==0.4.1
crispy-bootstrap5==0.7
cryptography==41.0.3
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.5
django-allauth==0.52.0
django-celery-results==2.5.1
django-core==1.4.1
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.11.2
django-storages==1.14
docutils==0.20.1
dulwich==0.21.6
extruct==0.16.0
filelock==3.12.3
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gunicorn==20.1.0
html-text==0.5.2
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
iniconfig==2.0.0
installer==0.7.0
isodate==0.6.1
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
jaraco.classes==3.3.0
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.17.3
jstyleson==0.0.2
keyring==24.2.0
kombu==5.3.2
loguru==0.7.1
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mf2py==1.1.3
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
parsel==1.8.1
pathspec==0.11.2
pdfkit==1.0.0
pdoc3==0.10.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.3.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.24.3
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyRdfa3==3.5.3
pyrsistent==0.19.3
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
queuelib==1.6.2
rapidfuzz==2.15.1
rdflib==7.0.0
readme-renderer==41.0
redis==5.0.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
rsa==4.9
ruff==0.0.269
scrapfly-sdk==0.8.8
Scrapy==2.10.1
SecretStorage==3.3.3
sentry-sdk==1.30.0
service-identity==23.1.0
shellingham==1.5.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==3.5.0
tomlkit==0.12.1
tqdm==4.66.1
trove-classifiers==2023.8.7
twine==4.0.2
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.4
w3lib==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
wkhtmltopdf==0.2
yarl==1.9.2
zipp==3.16.2
zope.interface==6.0
